CSV to BIN to XML Converter
In this assignment, you are asked to develop a command line tool to convert files between CSV, Binary, and XML file formats. In addition, to validate your XML file, a corresponding XSD file should be prepared. All codings must be in C programming language. Assignment will be done in groups of two.